Calculate the number of errors
The number of errors consists of the number of mismatches plus the
number of gap characters. These two quantities are calculated and made
accessible via getters Mismatches and Gaps. The number of errors is also
part of the revised header, where the descriptions of quantities now
distinguish between singular and plural.